[PPU] Reuse existing W8A16 benchmark for FP8
