+++ benchmark/test_fused_marlin_moe_w8a16_int8.py
@@ -37,10 +37,26 @@ except ImportError:
 import flaggems_vllm
 
 # FlagGems wrapper under test
-from flaggems_vllm.ops.fused_marlin_moe import QUANT_TYPE_UINT8B128
+from flaggems_vllm.ops.fused_marlin_moe import (
+    QUANT_TYPE_FP8_E4M3,
+    QUANT_TYPE_UINT8B128,
+)
 from flaggems_vllm.ops.fused_marlin_moe import fused_marlin_moe as gems_fused_marlin_moe
 
 from . import base
+
+try:
+    from vllm import _custom_ops as vllm_ops
+    from vllm.model_executor.layers.quantization.utils.marlin_utils import (
+        marlin_moe_permute_scales,
+    )
+    from vllm.model_executor.layers.quantization.utils.marlin_utils_fp8 import (
+        fp8_fused_exponent_bias_into_scales,
+    )
+
+    HAS_VLLM_FP8_MARLIN_MOE = HAS_VLLM_FUSED_MARLIN_MOE
+except ImportError:
+    HAS_VLLM_FP8_MARLIN_MOE = False
 
 
 def is_cuda_available():
@@ -253,4 +269,270 @@ def test_fused_marlin_moe_w8a16_int8():
         dtypes=[torch.bfloat16],
     )
     bench.set_gems(_gems_call_int8)
+    bench.run()
+
+
+PPU_FP8_MODEL_GEOMETRY = (256, 4096, 256, 6)
+PPU_FP8_PR5140_TRACE = (
+    (1, 172),
+    (2, 172),
+    (4, 172),
+    (8, 172),
+    (16, 172),
+    (24, 172),
+    (32, 172),
+    (40, 172),
+    (48, 172),
+    (56, 172),
+    (64, 172),
+    (72, 172),
+    (80, 172),
+    (88, 172),
+    (96, 172),
+    (104, 172),
+    (112, 172),
+    (120, 172),
+    (128, 172),
+    (136, 172),
+    (144, 172),
+    (152, 172),
+    (160, 172),
+    (168, 172),
+    (176, 172),
+    (184, 172),
+    (192, 172),
+    (200, 172),
+    (208, 172),
+    (216, 172),
+    (224, 172),
+    (232, 172),
+    (240, 172),
+    (248, 172),
+    (256, 172),
+    (272, 172),
+    (288, 172),
+    (304, 172),
+    (320, 172),
+    (336, 172),
+    (352, 172),
+    (368, 172),
+    (384, 172),
+    (400, 172),
+    (416, 172),
+    (432, 172),
+    (448, 172),
+    (464, 172),
+    (480, 172),
+    (496, 344),
+    (512, 344),
+    (2048, 43),
+    (16384, 946),
+)
+
+
+def _pack_gptq_int32_fp8(weight):
+    """Pack four E4M3 bytes per GPTQ INT32 word."""
+    experts, output_size, packed_k = weight.shape
+    packed = torch.empty(
+        (experts, packed_k // 4, output_size),
+        device=weight.device,
+        dtype=torch.int32,
+    )
+    for expert in range(experts):
+        bytes4 = weight[expert].to(torch.int32).reshape(output_size, packed_k // 4, 4)
+        words = (
+            bytes4[..., 0]
+            | (bytes4[..., 1] << 8)
+            | (bytes4[..., 2] << 16)
+            | (bytes4[..., 3] << 24)
+        )
+        packed[expert].copy_(words.transpose(0, 1))
+    return packed
+
+
+def _to_vllm_marlin_fp8(weight, scales, size_k, size_n):
+    qweight = _pack_gptq_int32_fp8(weight.view(torch.uint8))
+    empty_perm = torch.empty(
+        (weight.size(0), 0), device=weight.device, dtype=torch.int32
+    )
+    qweight = vllm_ops.gptq_marlin_moe_repack(
+        qweight,
+        empty_perm,
+        size_k=size_k,
+        size_n=size_n,
+        num_bits=8,
+    )
+    scales = marlin_moe_permute_scales(
+        scales.transpose(1, 2).contiguous(),
+        size_k=size_k,
+        size_n=size_n,
+        group_size=GROUP_SIZE,
+    )
+    return qweight, fp8_fused_exponent_bias_into_scales(scales)
+
+
+def _make_ppu_fp8_weights(num_experts, hidden_size, intermediate_size, dtype):
+    torch.manual_seed(7)
+    weights = []
+    scales = []
+    native_weights = []
+    native_scales = []
+    for output_size, input_size in (
+        (2 * intermediate_size, hidden_size),
+        (hidden_size, intermediate_size),
+    ):
+        raw = torch.randint(
+            0,
+            256,
+            (num_experts, output_size, input_size),
+            device=flaggems_vllm.device,
+            dtype=torch.uint8,
+        )
+        raw = torch.where((raw & 127) == 127, raw - 1, raw)
+        weight = raw.view(torch.float8_e4m3fn)
+        scale = (
+            torch.rand(
+                (num_experts, output_size, input_size // GROUP_SIZE),
+                device=flaggems_vllm.device,
+            )
+            * 0.001
+            + 0.0005
+        ).to(dtype)
+        native_weight, native_scale = _to_vllm_marlin_fp8(
+            weight, scale, input_size, output_size
+        )
+        weights.append(weight)
+        scales.append(scale)
+        native_weights.append(native_weight)
+        native_scales.append(native_scale)
+    return (*weights, *scales, *native_weights, *native_scales)
+
+
+class FusedMarlinMoEW8A16FP8Benchmark(base.Benchmark):
+    """Use the production trace from FlagGems PR 5140 on T-Head PPU."""
+
+    def set_shapes(self, shape_file_path=None):
+        num_experts, hidden_size, intermediate_size, top_k = PPU_FP8_MODEL_GEOMETRY
+        self.shapes = [
+            (m, num_experts, hidden_size, intermediate_size, top_k, count)
+            for m, count in PPU_FP8_PR5140_TRACE
+        ]
+        self.shape_desc = "M, E, K, N, top_k, call_count"
+
+    def get_input_iter(self, dtype):
+        num_experts, hidden_size, intermediate_size, top_k = PPU_FP8_MODEL_GEOMETRY
+        weights = _make_ppu_fp8_weights(
+            num_experts, hidden_size, intermediate_size, dtype
+        )
+        for num_tokens, call_count in PPU_FP8_PR5140_TRACE:
+            torch.manual_seed(7 + num_tokens)
+            hidden_states = (
+                torch.randn(
+                    (num_tokens, hidden_size),
+                    device=flaggems_vllm.device,
+                    dtype=dtype,
+                )
+                * 0.1
+            )
+            topk_ids = (
+                torch.rand((num_tokens, num_experts), device=flaggems_vllm.device)
+                .topk(top_k, dim=-1)
+                .indices
+            )
+            topk_weights = torch.softmax(
+                torch.randn((num_tokens, top_k), device=flaggems_vllm.device),
+                dim=-1,
+            ).to(torch.float32)
+            inputs = (
+                hidden_states,
+                *weights,
+                topk_weights,
+                topk_ids,
+                call_count,
+            )
+            actual = _gems_call_fp8(*inputs)
+            expected = _vllm_baseline_fp8(*inputs)
+            error = (actual.float() - expected.float()).abs().mean()
+            error /= expected.float().abs().mean().clamp_min(1e-12)
+            assert error < 0.04 and torch.isfinite(actual).all(), (
+                num_tokens,
+                error.item(),
+            )
+            yield inputs
+
+
+def _vllm_baseline_fp8(
+    hidden_states,
+    w1,
+    w2,
+    w1_scale,
+    w2_scale,
+    vllm_w1,
+    vllm_w2,
+    vllm_w1_scale,
+    vllm_w2_scale,
+    topk_weights,
+    topk_ids,
+    call_count,
+):
+    del w1, w2, w1_scale, w2_scale, call_count
+    return vllm_fused_marlin_moe(
+        hidden_states=hidden_states,
+        w1=vllm_w1,
+        w2=vllm_w2,
+        bias1=None,
+        bias2=None,
+        w1_scale=vllm_w1_scale,
+        w2_scale=vllm_w2_scale,
+        topk_weights=topk_weights,
+        topk_ids=topk_ids,
+        quant_type_id=scalar_types.float8_e4m3fn.id,
+    )
+
+
+def _gems_call_fp8(
+    hidden_states,
+    w1,
+    w2,
+    w1_scale,
+    w2_scale,
+    vllm_w1,
+    vllm_w2,
+    vllm_w1_scale,
+    vllm_w2_scale,
+    topk_weights,
+    topk_ids,
+    call_count,
+):
+    del vllm_w1, vllm_w2, vllm_w1_scale, vllm_w2_scale, call_count
+    return flaggems_vllm.fused_marlin_moe(
+        hidden_states,
+        w1,
+        w2,
+        None,
+        None,
+        w1_scale,
+        w2_scale,
+        topk_weights,
+        topk_ids,
+        QUANT_TYPE_FP8_E4M3,
+        group_size=GROUP_SIZE,
+    )
+
+
+@pytest.mark.fused_marlin_moe
+@pytest.mark.skipif(
+    flaggems_vllm.vendor_name != "thead", reason="T-Head PPU trace benchmark"
+)
+@pytest.mark.skipif(
+    not HAS_VLLM_FP8_MARLIN_MOE,
+    reason="vLLM FP8 fused_marlin_moe is not installed",
+)
+def test_fused_marlin_moe_w8a16_fp8():
+    bench = FusedMarlinMoEW8A16FP8Benchmark(
+        op_name="fused_marlin_moe_w8a16_fp8",
+        torch_op=_vllm_baseline_fp8,
+        dtypes=[torch.bfloat16],
+    )
+    bench.set_gems(_gems_call_fp8)
     bench.run()
